Write bulk file data to image
Bulk file data is split into individual chunks which
are enqueued for further processing. A queueing concept
ensures that processing can be delegated to multiple threads,
yet chunks written to the image in the order specified by
the front-end (needed to keep file data contiguous as
required by the format).