Write xattr table only when necessary
If no inode in the filesystem has extended
attributes, there is no point in creating an
xattr table. This results in a miniscule amount
of saved space, but more importantly it makes
the format bit-compatible with pre-xattr
squashfs format (mostly relevant for unsquashfs
and friends, as the kernel simply ignores
xattrs if they are unsupported).