Fix failure to seal image when no fragments present
The front end submits a final "closing piece" to the
worker threads that is intended to force all outstanding
fragments to disk. When there are no fragments present, this
closing piece was marked "finished" immediately (since there
is in fact nothing to do), but this causes an assertion
failure as it violates the invariant that outstanding
chunks must be finished strictly in the order they
were submitted.

While the assertion failure is harmless and could be
removed, fix this by marking the closing piece just
"writable" so the strict write-out order is maintained
even for the final "null" piece.