Introduce compress_meta method
Add method "compress_meta" (in addition to just "compress") to
compressor instances. This function is to be used for compression
of meta data blocks, change callers accordingly.

For most algorithms, the two functions do not differ in their
behaviour, but for xz it makes a difference. Route callees in
the different compressor implementations accordingly.

Signed-off-by: Helge Bahmann <helge.bahmann@secunet.com>